// client/src/generated-ts-client.ts
export interface ApplicationUserDto {
    id?: string;
    fullName?: string;
    email?: string;
    phoneNumber?: string;
    isActivePlayer?: boolean;
    role?: string;
    createdAt?: string;
    updatedAt?: string | undefined;
}

export interface RegisterPlayerRequest {
    email: string;
    fullName: string;
    phoneNumber: string;
}

export interface LoginResponse {
    token?: string;
    user?: ApplicationUserDto;
    role?: string;
    expires?: string;
}

export interface LoginRequest {
    email: string;
    password: string;
}